skills/shared/images-k-generation/scripts/pyproject.toml
[project]
name = "image-generation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "httpx>=0.28.1",
    "python-dotenv>=1.2.2",
]
